refactor(#8/*): domain 하위 패키지를 루트 하위로 이동
