--- src/main/java/org/sopt/domain/post/application/validator/PostValidationRule.java
@@ -1,5 +1,0 @@
-package org.sopt.domain.post.application.validator;
-
-public interface PostValidationRule {
-    void validate(Long userId, String title);
-}
