--- src/main/java/org/sopt/domain/post/application/validator/PostValidator.java
@@ -1,25 +1,0 @@
-package org.sopt.domain.post.application.validator;
-
-import java.util.List;
-import org.sopt.domain.post.application.validator.rule.CreatableTimeValidator;
-import org.sopt.domain.post.application.validator.rule.DuplicateTitleValidator;
-import org.sopt.domain.post.infrastructure.repository.PostRepository;
-import org.springframework.stereotype.Component;
-
-@Component
-public class PostValidator {
-    private final List<PostValidationRule> rules;
-
-    public PostValidator(PostRepository postRepository) {
-        this.rules = List.of(
-                new DuplicateTitleValidator(postRepository),
-                new CreatableTimeValidator(postRepository)
-        );
-    }
-
-    public void validateAll(Long userId, String title) {
-        for (PostValidationRule rule : rules) {
-            rule.validate(userId, title);
-        }
-    }
-}
