--- src/main/java/org/sopt/domain/post/presentation/response/GetAllPostsResponse.java
@@ -1,6 +1,0 @@
-package org.sopt.domain.post.presentation.response;
-
-import java.util.List;
-
-public record GetAllPostsResponse(List<GetSimplePostResponse> results) {
-}
