--- src/main/java/org/sopt/domain/post/presentation/response/GetDetailedPostResponse.java
@@ -1,15 +1,0 @@
-package org.sopt.domain.post.presentation.response;
-
-import org.sopt.domain.post.domain.Post;
-import org.sopt.domain.user.dto.response.GetUserResponse;
-
-public record GetDetailedPostResponse(Long id, GetUserResponse user, String title, String content) {
-    public static GetDetailedPostResponse from(Post post) {
-        return new GetDetailedPostResponse(
-                post.getId(),
-                GetUserResponse.from(post.getUser()),
-                post.getTitle(),
-                post.getContent()
-        );
-    }
-}
