--- src/main/java/org/sopt/domain/user/controller/UserController.java
@@ -1,32 +1,0 @@
-package org.sopt.domain.user.controller;
-
-import static org.sopt.domain.user.message.UserMessage.CREATED_SUCCESS;
-
-import jakarta.validation.Valid;
-import java.net.URI;
-import lombok.RequiredArgsConstructor;
-import org.sopt.global.response.ApiResponse;
-import org.sopt.domain.user.dto.request.CreateUserRequest;
-import org.sopt.domain.user.service.UserService;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RequestMapping("/api/v1/users")
-@RequiredArgsConstructor
-@RestController
-public class UserController {
-
-    private final UserService userService;
-
-    @PostMapping
-    public ResponseEntity<ApiResponse<Void>> signUp(@Valid @RequestBody CreateUserRequest createUserRequest) {
-        Long createdId = userService.createUser(createUserRequest);
-        URI location = URI.create("/api/v1/users/" + createdId);
-
-        return ResponseEntity.created(location).body(ApiResponse.created(CREATED_SUCCESS));
-    }
-}
