--- src/main/java/org/sopt/domain/user/message/UserMessage.java
@@ -1,5 +1,0 @@
-package org.sopt.domain.user.message;
-
-public class UserMessage {
-    public static final String CREATED_SUCCESS = "성공적으로 사용자를 등록했습니다.";
-}
