--- src/main/java/org/sopt/domain/user/repository/UserRepository.java
@@ -1,7 +1,0 @@
-package org.sopt.domain.user.repository;
-
-import org.sopt.domain.user.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-}
