+++ src/main/java/org/sopt/post/application/validator/PostValidationRule.java
@@ -1,0 +1,5 @@
+package org.sopt.post.application.validator;
+
+public interface PostValidationRule {
+    void validate(Long userId, String title);
+}
