+++ src/main/java/org/sopt/post/infrastructure/repository/PostRepository.java
@@ -1,0 +1,23 @@
+package org.sopt.post.infrastructure.repository;
+
+import java.util.List;
+import java.util.Optional;
+import org.sopt.post.domain.Post;
+import org.sopt.post.domain.Tag;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface PostRepository extends JpaRepository<Post, Long> {
+    Optional<Post> findFirstByUser_IdOrderByCreatedAtDesc(Long userId);
+
+    List<Post> findAllByOrderByCreatedAtDesc();
+
+    Optional<Post> findPostByTitle(String title);
+
+    List<Post> findPostsByTitleContaining(String keyword);
+
+    List<Post> findPostsByUser_nameContaining(String keyword);
+
+    List<Post> findPostsByTag(Tag tag);
+}
