+++ src/main/java/org/sopt/post/presentation/message/PostMessage.java
@@ -1,0 +1,11 @@
+package org.sopt.post.presentation.message;
+
+public class PostMessage {
+    public static final String CREATED_SUCCESS = "성공적으로 게시글을 등록했습니다.";
+    public static final String RETRIEVED_ALL_SUCCESS = "성공적으로 전체 게시글을 조회했습니다.";
+    public static final String SEARCHED_SUCCESS = "성공적으로 게시글을 검색했습니다.";
+    public static final String RETRIEVED_SUCCESS = "성공적으로 게시글을 조회했습니다.";
+    public static final String DELETED_SUCCESS = "성공적으로 게시글을 삭제했습니다.";
+    public static final String UPDATED_SUCCESS = "성공적으로 게시글을 수정했습니다.";
+}
+
