+++ src/main/java/org/sopt/post/presentation/response/GetAllPostsResponse.java
@@ -1,0 +1,6 @@
+package org.sopt.post.presentation.response;
+
+import java.util.List;
+
+public record GetAllPostsResponse(List<GetSimplePostResponse> results) {
+}
