+++ src/main/java/org/sopt/post/presentation/response/GetDetailedPostResponse.java
@@ -1,0 +1,15 @@
+package org.sopt.post.presentation.response;
+
+import org.sopt.post.domain.Post;
+import org.sopt.user.dto.response.GetUserResponse;
+
+public record GetDetailedPostResponse(Long id, GetUserResponse user, String title, String content) {
+    public static GetDetailedPostResponse from(Post post) {
+        return new GetDetailedPostResponse(
+                post.getId(),
+                GetUserResponse.from(post.getUser()),
+                post.getTitle(),
+                post.getContent()
+        );
+    }
+}
