+++ src/main/java/org/sopt/user/dto/response/GetUserResponse.java
@@ -1,0 +1,9 @@
+package org.sopt.user.dto.response;
+
+import org.sopt.user.domain.User;
+
+public record GetUserResponse(Long id, String name) {
+    public static GetUserResponse from(User user) {
+        return new GetUserResponse(user.getId(), user.getName());
+    }
+}
