+++ src/main/java/org/sopt/user/message/UserMessage.java
@@ -1,0 +1,5 @@
+package org.sopt.user.message;
+
+public class UserMessage {
+    public static final String CREATED_SUCCESS = "성공적으로 사용자를 등록했습니다.";
+}
