+++ src/main/java/org/sopt/user/repository/UserRepository.java
@@ -1,0 +1,7 @@
+package org.sopt.user.repository;
+
+import org.sopt.user.domain.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+}
